test(input): migrate legacy masking
